requirements.txt
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
openai>=1.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
